requirements.txt
# Core dependencies
google-generativeai>=0.3.0
openai>=1.0.0

# Optional but recommended
python-dotenv>=1.0.0
